Fix login flow
