+++ lib/features/root/launch_exception.dart
@@ -1,0 +1,12 @@
+class LaunchException {
+  final String message;
+  final Object underlyingException;
+
+  LaunchException(
+    this.message,
+    this.underlyingException,
+  );
+
+  @override
+  String toString() => message + underlyingException.toString();
+}
